requirements.txt
jinja2
progressbar2
wikipedia
wikiquote
